requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
SQLAlchemy==2.0.40
Werkzeug==3.1.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
email-validator==2.2.0
pandas==2.2.3
qrcode==8.1.0
openpyxl==3.1.5
Pillow==11.2.1
WTForms==3.2.1
gunicorn==23.0.0 